requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.0.2
Flask-SocketIO==3.3.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
monotonic==1.5
python-engineio==3.5.1
python-firebase==1.2
python-socketio==3.1.2
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.2
